let EsriShapefile handle an unknown type
